chore: Address reviewer concern
